weather_forecasts/requirements.txt
google-cloud-storage
pandas
requests
beautifulsoup4